Refactor to use mqtt internally for queueing messages.
I wanted to make the messagebus more robust and resilient
to failures by adding a queueing mechanism. MQTT is already
popular for IoT, so it seemed like a natural choice.

I broke up the application into an HTTP frontend that
writes events to MQTT and a backend that reads from MQTT
and dispatches to the subscribers.

Also added a form for submitting simple heartbeats that
only requires a device_id.